+++ REngine.RPI.RenderGraph/IResourceManager.cs
@@ -9,16 +9,16 @@ namespace REngine.RPI.RenderGraph
 {
 	public struct ResourceInsertInfo
 	{
-		public int Id;
+		public ulong Id;
 		public IGPUObject Resource;
 	}
 	public interface IResourceManager
 	{
 		public IResource GetResource(string resourceName);
-		public IResource GetResource(int resourceId);
+		public IResource GetResource(ulong resourceId);
 		public IResourceManager AddResource(in ResourceInsertInfo resource);
 		public IResourceManager AddResources(IEnumerable<ResourceInsertInfo> resources);
 		public IResourceManager UpdateResource(string resourceName, IGPUObject resource);
-		public IResourceManager UpdateResource(int resourceId, IGPUObject resource);
+		public IResourceManager UpdateResource(ulong resourceId, IGPUObject resource);
 	}
 }
